Converter SPI type (int) matches to current JS version
