+++ platform/dolphin-platform-core/src/main/java/com/canoo/dolphin/impl/BeanRepositoryImpl.java
@@ -73,7 +73,7 @@ public class BeanRepositoryImpl implements BeanRepository{
 
     @Override
     public <T> Subscription addOnAddedListener(final Class<T> beanClass, final BeanAddedListener<? super T> listener) {
-        BeanUtils.checkClass(beanClass);
+        DolphinUtils.assertIsDolphinBean(beanClass);
         beanAddedListenerMap.put(beanClass, listener);
         return new Subscription() {
             @Override
@@ -96,7 +96,7 @@ public class BeanRepositoryImpl implements BeanRepository{
 
     @Override
     public <T> Subscription addOnRemovedListener(final Class<T> beanClass, final BeanRemovedListener<? super T> listener) {
-        BeanUtils.checkClass(beanClass);
+        DolphinUtils.assertIsDolphinBean(beanClass);
         beanRemovedListenerMap.put(beanClass, listener);
         return new Subscription() {
             @Override
@@ -119,13 +119,13 @@ public class BeanRepositoryImpl implements BeanRepository{
 
     @Override
     public boolean isManaged(Object bean) {
-        BeanUtils.checkBean(bean);
+        DolphinUtils.assertIsDolphinBean(bean);
         return objectPmToDolphinPm.containsKey(bean);
     }
 
     @Override
     public <T> void delete(T bean) {
-        BeanUtils.checkBean(bean);
+        DolphinUtils.assertIsDolphinBean(bean);
         final PresentationModel model = objectPmToDolphinPm.remove(bean);
         if (model != null) {
             dolphinIdToObjectPm.remove(model.getId());
@@ -136,7 +136,7 @@ public class BeanRepositoryImpl implements BeanRepository{
     @Override
     @SuppressWarnings("unchecked")
     public <T> List<T> findAll(Class<T> beanClass) {
-        BeanUtils.checkClass(beanClass);
+        DolphinUtils.assertIsDolphinBean(beanClass);
         final List<T> result = new ArrayList<>();
         final List<PresentationModel> presentationModels = dolphin.findAllPresentationModelsByType(DolphinUtils.getDolphinPresentationModelTypeForClass(beanClass));
         for (PresentationModel model : presentationModels) {
@@ -155,7 +155,7 @@ public class BeanRepositoryImpl implements BeanRepository{
         if (bean == null) {
             return null;
         }
-        BeanUtils.checkBean(bean);
+        DolphinUtils.assertIsDolphinBean(bean);
         try {
             return objectPmToDolphinPm.get(bean).getId();
         } catch (NullPointerException ex) {
@@ -166,7 +166,7 @@ public class BeanRepositoryImpl implements BeanRepository{
     @Override
     @SuppressWarnings("unchecked")
     public void registerBean(Object bean, PresentationModel model, UpdateSource source) {
-        BeanUtils.checkBean(bean);
+        DolphinUtils.assertIsDolphinBean(bean);
         objectPmToDolphinPm.put(bean, model);
         dolphinIdToObjectPm.put(model.getId(), bean);
