+++ platform/dolphin-platform-core/src/main/java/com/canoo/dolphin/impl/ConverterFactory.java
@@ -8,7 +8,7 @@ public interface ConverterFactory {
 
     boolean supportsType(Class<?> cls);
 
-    String getTypeIdentifier();
+    int getTypeIdentifier();
 
     Converter getConverterForType(Class<?> cls);
