--- platform/dolphin-platform-core/src/main/java/com/canoo/dolphin/impl/converters/DateConverter.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2015-2016 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.canoo.dolphin.impl.converters;
-
-import com.canoo.dolphin.impl.Converter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.TimeZone;
-
-public class DateConverter implements Converter {
-
-    private static final Logger LOG = LoggerFactory.getLogger(DateConverter.class);
-
-    private final DateFormat dateFormat;
-
-    public DateConverter() {
-        dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
-        dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-    }
-
-    @Override
-    public Object convertFromDolphin(Object value) {
-        if (value == null) {
-            return null;
-        }
-        try {
-            return dateFormat.parse(value.toString());
-        } catch (ParseException e) {
-            LOG.warn("Unable to parse the date: " + value);
-            return null;
-        }
-    }
-
-    @Override
-    public Object convertToDolphin(Object value) {
-        if (value == null) {
-            return null;
-        }
-        try {
-            return dateFormat.format(value);
-        } catch (IllegalArgumentException ex) {
-            LOG.warn("Unable to format the date: " + value);
-            return null;
-        }
-    }
-}
