+++ platform/dolphin-platform-core/src/main/java/com/canoo/dolphin/impl/converters/EnumConverterFactory.java
@@ -1,13 +1,15 @@
 package com.canoo.dolphin.impl.converters;
 
 import com.canoo.dolphin.impl.Converter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class EnumConverterFactory extends AbstractConverterFactory {
 
-    public final static String FIELD_TYPE_ENUM = "E";
+    public final static int FIELD_TYPE_ENUM = 10;
 
     private final Map<Class<?>, EnumConverter> enumConverters = new HashMap<>();
 
@@ -17,7 +19,7 @@ public class EnumConverterFactory extends AbstractConverterFactory {
     }
 
     @Override
-    public String getTypeIdentifier() {
+    public int getTypeIdentifier() {
         return FIELD_TYPE_ENUM;
     }
 
@@ -30,4 +32,43 @@ public class EnumConverterFactory extends AbstractConverterFactory {
         }
         return enumConverter;
     }
+
+    private static class EnumConverter implements Converter {
+
+        private static final Logger LOG = LoggerFactory.getLogger(EnumConverter.class);
+
+        private final Class<? extends Enum> clazz;
+
+        @SuppressWarnings("unchecked")
+        public EnumConverter(Class<?> clazz) {
+            this.clazz = (Class<? extends Enum>) clazz;
+        }
+
+        @Override
+        public Object convertFromDolphin(Object value) {
+            if (value == null) {
+                return null;
+            }
+            try {
+                return Enum.valueOf(clazz, value.toString());
+            } catch (IllegalArgumentException ex) {
+                LOG.warn("Unable to convert to an enum (%s): %s", clazz, value);
+                return null;
+            }
+        }
+
+        @Override
+        public Object convertToDolphin(Object value) {
+            if (value == null) {
+                return null;
+            }
+            try {
+                return ((Enum)value).name();
+            } catch (ClassCastException ex) {
+                LOG.warn("Unable to evaluate the enum: " + value);
+                return null;
+            }
+        }
+    }
+
 }
